fixed GET_UINT32 was broken
git-svn-id: https://svn.qrtech.se/svn/Sources/SM10002/trunk@23690 6bce8083-c9ef-d345-8e70-90becfb1c772

+++ source/Xcp_ByteStream.h
@@ -21,7 +21,7 @@
 
 #define GET_UINT8(data, offset)  (*((uint8* )(data)+(offset)))
 #define GET_UINT16(data, offset) (*(uint16*)((uint8*)(data)+(offset)))
-#define GET_UINT32(data, offset) (*(uint16*)((uint8*)(data)+(offset)))
+#define GET_UINT32(data, offset) (*(uint32*)((uint8*)(data)+(offset)))
 
 
 #define SET_UINT8(data, offset, value) do {                 \
